feat: Adjust left page margin to fake align table on perceived margin
Reduces left page margin by the width of the first column and then pads all other major elemnts by that size to account for page margin change.

This is a hacky way of moving the table left.
An alternative nicer approach of wrapping it in a `pad()` function instead didn't work as it breaks page wrapping for the table.